refactored: proposals page
