+++ src/screens/proposals/index.tsx
@@ -5,10 +5,17 @@ import {
 } from '@components';
 import { useStyles } from './styles';
 import { List } from './components';
+import { useProposals } from './hooks';
 
 const Proposals = () => {
   const { t } = useTranslation('proposals');
   const classes = useStyles();
+  const {
+    state,
+    loadMoreItems,
+    itemCount,
+    isItemLoaded,
+  } = useProposals();
 
   return (
     <Layout
@@ -16,7 +23,13 @@ const Proposals = () => {
       navTitle={t('proposals')}
       className={classes.root}
     >
-      <List />
+      <List
+        items={state.items}
+        rawDataTotal={state.rawDataTotal}
+        isItemLoaded={isItemLoaded}
+        itemCount={itemCount}
+        loadMoreItems={loadMoreItems}
+      />
     </Layout>
   );
 };
