+++ src/screens/proposals/types.ts
@@ -1,0 +1,15 @@
+export type ProposalType = {
+  id: number;
+  title: string;
+  description: string;
+  status: string;
+}
+
+export type ProposalsState = {
+  loading: boolean;
+  exists: boolean;
+  hasNextPage: boolean;
+  isNextPageLoading: boolean;
+  rawDataTotal: number;
+  items: ProposalType[];
+}
